리뷰 통신 하는 중

+++ Travely_User/app/src/main/java/com/tooc/android/tooc/model/store/ReviewResponseData.kt
@@ -2,8 +2,8 @@ package com.tooc.android.tooc.model.store
 
 data class ReviewResponseData (
         var content : String,
-        var createdAt : Double,
-        var like : Int,
+        var createdAt : Long,
+        var like : Double,
         var reviewIdx : Int,
         var storeIdx : Int,
         var userIdx : Int,
